% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcs_list_mountpoints.R
\name{gcs_list_mountpoints}
\alias{gcs_list_mountpoints}
\title{List all GCS mountpoints}
\usage{
gcs_list_mountpoints()
}
\value{
a data.frame object
}
\description{
List all GCS mountpoints
}
\examples{
gcs_list_mountpoints()
}
