% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcs_mount.R
\name{gcs_mount}
\alias{gcs_mount}
\title{Mount a Google Cloud Storage(GCS) file system}
\usage{
gcs_mount(
  remote,
  mountpoint,
  mode = c("r", "rw"),
  cache_type = c("disk", "memory", "none"),
  cache_arg = NULL,
  billing = NULL,
  refresh = 60,
  implicit_dirs = TRUE,
  key_file = NULL,
  additional_args = NULL
)
}
\arguments{
\item{remote}{the remote path to a Google Cloud Storage bucket.
It can be either the bucket itself or a directory in the bucket.}

\item{mountpoint}{The mount point where the GCS file system will
be mounted to.}

\item{mode}{the permission of the mounted directory. The write
permission is only available on Linux.}

\item{cache_type}{The location where the file cache will be
stored. The cache type `none` and `memory` are only available
on Windows}

\item{cache_arg}{The argument of a cache type. If 
`cache_type = "memory"`, the argument is the limit of
the memory useage in MB. If `cache_type = "disk"`, the 
argument is the path to a cache directory.}

\item{billing}{The billing project ID.}

\item{refresh}{The refresh rate of the cloud files.}

\item{implicit_dirs}{Implicitly define directories based on content.
This argument is only available on Linux.}

\item{key_file}{The service account credentials file}

\item{additional_args}{The additional argument that will be passed to
the command-line program.}
}
\description{
The function uses the command-line program `GCSDokan` 
on Windows or `gcsfuse` on Linux system to mount 
a google cloud bucket path to your local file system.
}
