% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcs_unmount.R
\name{gcs_unmount}
\alias{gcs_unmount}
\title{Unmound a mounted GCS file system}
\usage{
gcs_unmount(mountpoint)
}
\arguments{
\item{mountpoint}{The path to the mounted GCS file system,
the value must match the one in `gcs_list_mountpoints`.}
}
\value{
No return value
}
\description{
Unmound a mounted GCS file system
}
\examples{
## Unmount the Z driver letter
## It has no effect if Z driver is not 
## a GCS mount point
gcs_unmount("Z")
}
