% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcs_list_mountpoints.R
\name{gcs_list_mountpoints}
\alias{gcs_list_mountpoints}
\title{List all GCS mountpoints}
\usage{
gcs_list_mountpoints()
}
\value{
a data.frame object with the first column named `remote`
and second named `mountpoint`
}
\description{
List all GCS mountpoints. The function uses `GCSDokan``
on Windows or the command `df` on Linux to show all 
mountpoints. Due to the system differences, the function
is only able to show the bucket name on Linux and can
show the full remote path on Windows.
}
\examples{
gcs_list_mountpoints()
}
