% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcs_unmount.R
\name{gcs_unmount}
\alias{gcs_unmount}
\title{Unmound a mounted GCS file system}
\usage{
gcs_unmount(mountpoint)
}
\arguments{
\item{mountpoint}{The path to the mounted GCS file system.}
}
\value{
No return value
}
\description{
Unmound a mounted GCS file system
}
\examples{
## Unmount a directory
## No operation will be done 
## if the directory does not exist or
## not a mounted directory
gcs_unmount("path_to_your_mounted_directory")
}
