% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcs_unmount.R
\name{gcs_unmount}
\alias{gcs_unmount}
\title{Unmound a mounted GCS file system}
\usage{
gcs_unmount(mountpoint)
}
\arguments{
\item{mountpoint}{The path to the mounted GCS file system.}
}
\value{
No return value
}
\description{
Unmound a mounted GCS file system
}
\examples{
## Unmount the Z driver letter
## Z driver must be a GCS mount point
gcs_unmount("Z")
}
